pyproject.toml
[project]
name = "gitbacklist"
version = "0.1.0"
description = "Incremental git backup tool for GitHub profiles and arbitrary repositories"
authors = [
    {name = "Lunar", email = "lunarlifeburst@gmail.com"},
]
dependencies = [
    "PyGithub>=2.1",
]
requires-python = ">=3.9"
readme = "README.md"
license = {text = "MIT"}

[build-system]
requires = ["pdm-backend"]
build-backend = "pdm.backend"


[tool.pdm]
distribution = true
